+++ admin/admin.php
@@ -71,7 +71,7 @@ if ($_SESSION['nick'] != 'admin') {
             </div>
         </section>
     </main>
-    <?php include "../footer.php" ?>
+    <?php include "footer.php" ?>
     <script src="assets/bootstrap/js/bootstrap.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/baguettebox.js/1.10.0/baguetteBox.min.js"></script>
     <script src="assets/js/smoothproducts.min.js"></script>