+++ CHANGELOG.md
@@ -14,9 +14,12 @@
 
 ### Nuevos 
 
-* [*] Creado archivo header-admin-logged.php-> Cabecera para el administrador
+* [*] Creado header-admin-logged.php        -> Cabecera para el administrador
 * [*] Creado archivo admin/admin.php        -> Página con enlaces para administrar entidades
-* [+] Creado archivo /admin/tiposanuncio.php-> CRUD completo de administración de tipos de anuncio
+* [+] Creado /admin/tiposanuncio.php        -> CRUD completo de administración de tipos de anuncio
+* [+] Creado /admin/tiposvivienda.php       -> CRUD completo de administración de tipos de vivienda
+* [+] Creado /admin/footer.php              -> Pie de administración (igual que footer.php)
+                                            Cambia las rutas relativas
 
 ### Cambios